Move helper functions to a common test helpers module
